Remove requirement for all args on construction
There are a lot of cases where this may just be used as a lib to
interface with the API (without registering). This update removes
the requirement for the positional args up front, and defaults them
to None instead. Users who need to register should obv still provide
them (non-breaking, they are still positional)